requirements.txt
cloudpickle==1.6.0
contourpy==1.3.0
cycler==0.12.1
fonttools==4.59.2
future==1.0.0
gym==0.19.0
importlib_resources==6.5.2
kiwisolver==1.4.7
llvmlite==0.43.0
matplotlib==3.9.4
numba==0.60.0
numpy==1.26.4
opencv-python==4.12.0.88
packaging==25.0
pillow==11.3.0
pygame==2.6.1
pyglet==1.5.27
PyOpenGL==3.1.10
PyOpenGL-accelerate==3.1.10
pyparsing==3.2.3
python-dateutil==2.9.0.post0
PyYAML==6.0.2
scipy==1.12.0
six==1.17.0
zipp==3.23.0
